+++ witherable/CHANGELOG.md
@@ -1,3 +1,10 @@
+0.4.1
+-------
+* Added `ordNubBy`, `hashNubBy`, `ordNubByOf`, and `hashNubByOf`.
+* Use `alterF` for nub-function implementations
+* Implement `witherM` in `Witherable Vector` instance.
+* Mark modules as Trustworthy
+
 0.4
 -------
 * `FilterableWithIndex` and `WitherableWithIndex` are now subclasses of the ones from [indexed-traversable](https://hackage.haskell.org/package/indexed-traversable)